requirements.txt
azure-functions
azure-cosmos
openai
pandas
python-dotenv
requests
